misc changes added (see description)
1. another good hashing code added
2. knapsack 1 and 2 added
3. nCr calculating tabulation method added
4. prefix function from cp-algorithms added
5. segment tree update version bug fixed
6. cses tree algorithms problems (tree diameter and tree distances II) solution added

+++ code-library/segment_tree_code.cpp
@@ -22,7 +22,7 @@ void segBuild(int node, int s, int e)
     tree[node] = tree[left] + tree[right];
 }
 
-void segUpdate(int node, int s, int e, int l, int r, int val)
+void segUpdate(int node, int s, int e, int l, int r, ll val)
 {
     int mid = (s + e) / 2;
     int left = 2 * node;
